backend/requirements.txt
# Core Framework
fastapi>=0.128.0
uvicorn[standard]>=0.34.0

# Database
sqlmodel>=0.0.31
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10

# Validation & Settings
pydantic>=2.12.5
python-dotenv>=1.2.1

# Authentication
pyjwt>=2.8.0
bcrypt>=4.0.1

# API Utilities
python-multipart>=0.0.20
slowapi>=0.1.9
websockets>=13.1

# AI/LLM
openai>=1.57.4
openai-agents>=0.6.5

# Date/Time utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Task Queue (required for reminders & recurring tasks)
celery>=5.3.0
redis>=4.5.0

# Code Linting (for CI/CD)
ruff>=0.1.6
black>=23.12.0
flake8>=7.0.0